feat: allow parsing of expanded geometries
Based on changes upstream in seq_geom_parser, this
allows transforming geometries that have unbounded
segments followed by fixed length segments, as long
as the fixed length segments are followed by an
anchor.

Also, this commit fixes a bug where the shortcut
condition in parse_single_read for returning the
whole string was wrong. The special case that anticipated
that the *whole string* matched the regex was wrong
on multiple levels.  The solution was to simply
eliminate that special case.